Implement rewrite-based anti-unification
We find the intersection between multiple terms using a Dfta,
then use it to enumerate all anti-unified possibilities in the
Dfta and then build patterns which are used as Searchers and
Appliers in rewrite rules, which we apply to an EGraph.

For performance, on a small example, Dfta creation takes about
2% of time (~40 us), program enumeration takes about 20% of time
(~400 us), and applying the rewrites takes the rest of the time
(78%, ~1600 us). Parallelizing program enumeration could yield
a nontrivial speedup, but most of the opportunities for improving
performance will probably come from getting a better algorithm
to perform anti-unification - perhaps a custom Searcher that
doesn't require us to explicitly enumerate every program (look
into egg::Searcher, egg::SearchMatches), or a custom Applier
to go along with that.