PoC rewrite extraction
It's in a very broken state right now. The actual rewrites are
commented out for testing purposes, to test which rewrites work
and which don't. Rewrite pattern matching is extremely
finicking, implying that either the Pattern Searcher is buggy
or that we need to switch to modifying the egraph directly
ourselves.