Add wip separate extraction impl
This implementation calculates an EnvCost for every eclass in the
egraph, using multiple passes to account for recursive cases.
Each EnvCost contains a map from a function environment to a cost
given that function environment. If an expression in that eclass
exists with a larger cost than an entry in the EnvCost with a subset
of the environment, that expression's cost and env isn't added.