change runner to take list of exprs as input
this change makes it so that the runner doesn't need a single huge
pre-combined program containing all sub-programs as an input for
anti-unification. we did this because anti-unifying dreamcoder examples
had a huge perf penalty with the root combine node already included in
our anti-unif target.

this change should restore dreamcoder performance to how it was before,
since it always had anti-unif operate on individual programs before
adding the root node back in.

note that this change doesn't change any of the other example files
except for the psych examples - keep this in mind if we get weird
behavior in the future :))